extract the `implied_bounds` code into a helper function
