+++ src/librustc/infer/outlives/mod.rs
@@ -8,5 +8,8 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+//! Various code related to computing outlives relations.
+
 pub mod env;
+pub mod implied_bounds;
 mod obligations;
